Updates Backend and Frontend code
This commit includes updates to the Backend and Frontend code. The changes consist of adding a new `.gitignore` file to the Backend, as well as adding multiple new Excel files to the `masivo` directory. Additionally, the commit includes modifications to the `facturasC.routes.js` and `masivo.routes.js` files in the Backend, as well as changes to the `facturas.component.html`, `facturas.component.ts`, `flujo-de-caja.component.html`, `form-consolidacion.component.ts`, `auth.service.ts`, and `consolidado.service.ts` files in the Frontend.

Fixes #789